requirements.txt
streamlit==1.31.0
numpy==1.26.3
opencv-python==4.9.0.80
Pillow==10.2.0
scipy==1.12.0
